befor marge
